Calculator branch merging

+++ src/main/java/Student/Calculator.java
@@ -1,16 +1,16 @@
 package Student;
 
 public class Calculator {
-    public double Add(float a,float b){
+    public float Add(float a,float b){
         return  a+b;
     }
-    public double Sub(float a,float b){
+    public float Sub(float a,float b){
         return a-b;
     }
-    public double Mul(float a,float b){
+    public float Mul(float a,float b){
         return a*b;
     }
-    public double Div(float a,float b){
+    public float Div(float a,float b){
         if(b==0){
             throw new ArithmeticException("Cannot divide by zero");
         }
